+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.5 (unreleased)
 
 - Added support for `halfvec`, `bit`, and `sparsevec` types
+- Added `registerTypes` function
 
 ## 0.1.4 (2023-12-08)